Refactoring widgets
