--- gaze_controlled_cursor_demo/__main__.py
@@ -1,3 +1,0 @@
-import app
-
-app.run()
