+++ gaze_controlled_cursor_demo/widgets/debug_window.py
