+++ gaze_controlled_cursor_demo/widgets/gaze_overlay.py
